requirements.txt
# Global International Relations Monitor V2
# Event-Centric Analysis Pipeline Dependencies

# Core HTTP and Parsing
requests>=2.28.0
beautifulsoup4>=4.11.0
feedparser>=6.0.0
newspaper3k>=0.2.8
python-dateutil>=2.8.0
lxml_html_clean>=0.1.0

# Database
psycopg2-binary>=2.9.0

# Environment Configuration
python-dotenv>=1.0.0

# OpenAI API
openai>=1.0.0

# API Server
flask>=3.0.0
gunicorn>=21.0.0

# Translation (MarianMT via Hugging Face)
transformers>=4.36.0
torch>=2.1.0
sentencepiece>=0.1.99

# Language Detection & NLP
langdetect>=1.0.9
spacy>=3.7.0
